+++ src/oop/CMakeLists.txt
@@ -116,9 +116,10 @@ catkin_package(
 ## Specify additional locations of header files
 ## Your package locations should be listed before other locations
 include_directories(
-# include
+  include
   ${catkin_INCLUDE_DIRS}
 )
+add_library(chassis src/chassis.cpp)
 
 ## Declare a C++ library
 # add_library(${PROJECT_NAME}
@@ -137,6 +138,9 @@ include_directories(
 
 add_executable(oop_ex1 src/oop_ex1.cpp)
 target_link_libraries(oop_ex1 ${catkin_LIBRARIES})
+
+add_executable(oop_ex2 src/oop_ex2.cpp)
+target_link_libraries(oop_ex2 ${catkin_LIBRARIES})
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
